refactor(chatpanel): split cardParsers into per-card modules
Move the 714-line helpers/cardParsers.ts into helpers/cardParsers/ with
one module per card kind (file, website, workItem, command, agentMessage,
orgtrackEnvelope) plus shared primitives; cardParsers/index.ts re-exports
the same 7 parsers + OrgtrackEnvelopeContext, so "./cardParsers" still
resolves for all 6 import sites. Pure code motion, no behaviour change.
