+++ src/engines/ChatPanel/blocks/ToolCallBlock/helpers.ts
@@ -5,7 +5,7 @@
  * Source files:
  *   helpers/argsSummary.ts   — screenshot utilities + extractArgsSummary
  *   helpers/resultParsers.ts — extractResultText, parseSearchFiles, …
- *   helpers/cardParsers.ts   — parseFileCard, parseWebsiteCard, …
+ *   helpers/cardParsers/     — parseFileCard, parseWebsiteCard, …
  */
 export {
   hasStyledOutput,
